+++ src/nidas/dynld/raf/GPS_HW_HG2021GB02.h
@@ -48,7 +48,7 @@ public:
   }
 
   /** Process all labels from this instrument. */
-  float processLabel(const long data);
+  float processLabel(const int data);
 
  private:
